# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_coindistro_backend_internal_earn_models.CompleteLearnRequest:
    properties:
      metadata:
        additionalProperties: true
        description: Optional metadata from Academy client
        type: object
    type: object
  github_com_coindistro_backend_internal_earn_models.CreateLaunchpoolRequest:
    properties:
      allocation_rules:
        additionalProperties: true
        type: object
      description:
        type: string
      name:
        type: string
      product_id:
        type: string
      reward_distribution:
        additionalProperties: true
        type: object
      status:
        type: string
      supported_assets:
        items:
          type: string
        minItems: 1
        type: array
      window_end:
        type: string
      window_start:
        type: string
    required:
    - name
    - supported_assets
    - window_end
    - window_start
    type: object
  github_com_coindistro_backend_internal_earn_models.CreateLearnCampaignRequest:
    properties:
      academy_course_id:
        type: string
      description:
        type: string
      ends_at:
        type: string
      name:
        type: string
      product_id:
        type: string
      reward_amount:
        type: number
      reward_asset:
        type: string
      starts_at:
        type: string
      status:
        type: string
    required:
    - name
    type: object
  github_com_coindistro_backend_internal_earn_models.CreateProductRequest:
    properties:
      capacity_total:
        type: number
      category:
        type: string
      description:
        type: string
      duration_days:
        type: integer
      eligibility:
        additionalProperties: true
        type: object
      ends_at:
        type: string
      featured:
        type: boolean
      max_allocation:
        type: number
      metadata:
        additionalProperties: true
        type: object
      min_allocation:
        type: number
      name:
        maxLength: 255
        minLength: 2
        type: string
      reward_apr:
        type: number
      reward_model:
        type: string
      risk_level:
        type: string
      rules:
        additionalProperties: true
        type: object
      slug:
        maxLength: 255
        minLength: 2
        type: string
      starts_at:
        type: string
      status:
        type: string
      strategy_profiles:
        items:
          type: string
        type: array
      supported_assets:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - category
    - name
    - slug
    - supported_assets
    type: object
...
  github_com_coindistro_backend_internal_earn_models.ProductAnalytics:
    properties:
      avg_allocation:
        type: number
      capacity_total:
        type: number
      capacity_used:
        type: number
      capacity_used_pct:
        type: number
      participants:
        type: integer
      product_id:
        type: string
      total_allocated:
        type: number
      total_rewards:
        type: number
    type: object
...
  github_com_coindistro_backend_internal_response.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIError'
      message:
        type: string
      meta:
        $ref: '#/definitions/github_com_coindistro_backend_internal_response.Meta'
      success:
        type: boolean
    type: object
...
host: coindistro.onrender.com
info:
  contact:
    email: support@coindistro.com
    name: Coindistro Support
    url: https://coindistro.com/support
  description: Coindistro — One Platform. Everything Crypto. API backend for Africa's
    next-generation crypto financial ecosystem.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://coindistro.com/terms
  title: Coindistro API
  version: 1.0.0
paths:
...
  /earn/admin/launchpool:
    post:
      consumes:
      - application/json
      parameters:
      - description: Campaign
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_coindistro_backend_internal_earn_models.CreateLaunchpoolRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Create launchpool campaign (admin)
      tags:
      - Earn Admin
  /earn/admin/learn:
    post:
      consumes:
      - application/json
      parameters:
      - description: Campaign
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_coindistro_backend_internal_earn_models.CreateLearnCampaignRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Create learn & earn campaign (admin)
      tags:
      - Earn Admin
  /earn/admin/products:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
      security:
      - BearerAuth: []
      summary: List all products including drafts (admin)
      tags:
      - Earn Admin
    post:
      consumes:
      - application/json
      parameters:
      - description: Product
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_coindistro_backend_internal_earn_models.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Create earn product (admin)
      tags:
      - Earn Admin
  /earn/admin/products/{id}:
    put:
      consumes:
      - application/json
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Updates
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_coindistro_backend_internal_earn_models.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update earn product (admin)
      tags:
      - Earn Admin
  /earn/admin/products/{id}/analytics:
    get:
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_coindistro_backend_internal_earn_models.ProductAnalytics'
              type: object
      security:
      - BearerAuth: []
      summary: Product analytics (admin)
      tags:
      - Earn Admin
  /earn/admin/products/{id}/participants:
    get:
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
      security:
      - BearerAuth: []
      summary: List product participants (admin)
      tags:
      - Earn Admin
  /earn/admin/products/{id}/status:
    put:
      consumes:
      - application/json
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Status payload
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Pause/resume/archive product (admin)
      tags:
      - Earn Admin
  /earn/history:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
      security:
      - BearerAuth: []
      summary: List earn transaction history
      tags:
      - Earn
...
  /earn/learn/{id}/complete:
    post:
      consumes:
      - application/json
      parameters:
      - description: Campaign ID
        in: path
        name: id
        required: true
        type: string
      - description: Metadata
        in: body
        name: body
        schema:
          $ref: '#/definitions/github_com_coindistro_backend_internal_earn_models.CompleteLearnRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Record learning completion for reward eligibility
      tags:
      - Earn
  /earn/participations:
    get:
      parameters:
      - description: Status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
      security:
      - BearerAuth: []
      summary: List user participations
      tags:
      - Earn
  /earn/participations/{id}:
    get:
      parameters:
      - description: Participation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get participation details
      tags:
      - Earn
  /earn/participations/{id}/add-funds:
    post:
      consumes:
      - application/json
      parameters:
      - description: Participation ID
        in: path
        name: id
        required: true
        type: string
      - description: Amount
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_coindistro_backend_internal_earn_models.AddFundsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Add funds to a participation
      tags:
      - Earn
  /earn/participations/{id}/exit:
    post:
      consumes:
      - application/json
      parameters:
      - description: Participation ID
        in: path
        name: id
        required: true
        type: string
      - description: Reason
        in: body
        name: body
        schema:
          $ref: '#/definitions/github_com_coindistro_backend_internal_earn_models.ExitParticipationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Exit a participation
      tags:
      - Earn
  /earn/participations/{id}/withdraw:
    post:
      consumes:
      - application/json
      parameters:
      - description: Participation ID
        in: path
        name: id
        required: true
        type: string
      - description: Amount
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_coindistro_backend_internal_earn_models.WithdrawRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Withdraw from a participation
      tags:
      - Earn
  /earn/portfolio:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_coindistro_backend_internal_earn_models.PortfolioOverview'
              type: object
      security:
      - BearerAuth: []
      summary: Earn portfolio overview
      tags:
      - Earn
  /earn/products:
    get:
      parameters:
      - description: Product category
        in: query
        name: category
        type: string
      - description: Status filter
        in: query
        name: status
        type: string
      - description: Featured only
        in: query
        name: featured
        type: boolean
      - description: Supported asset
        in: query
        name: asset
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
      summary: List earn products
      tags:
      - Earn
  /earn/products/{id}:
    get:
      parameters:
      - description: Product ID or slug
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
      summary: Get earn product by ID or slug
      tags:
      - Earn
  /earn/products/{id}/join:
    post:
      consumes:
      - application/json
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Join request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_coindistro_backend_internal_earn_models.JoinProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Join an earn product
      tags:
      - Earn
  /earn/referral/rewards:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_coindistro_backend_internal_earn_models.ReferralRewardSummary'
              type: object
      security:
      - BearerAuth: []
      summary: Referral reward summary and history
      tags:
      - Earn
  /earn/rewards:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
      security:
      - BearerAuth: []
      summary: List reward history
      tags:
      - Earn
...
schemes:
- https
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: Enter the token with the `Bearer ` prefix, e.g. "Bearer abcde12345".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"